#1694 - Improve annotation detail sidebar architecture and performance
- Merge rest of AnnotationEditorPanel (was AnnotationFeatureForm ) into AnnotationDetailEditorPanel
