+++ webanno-ui-annotation/src/main/java/de/tudarmstadt/ukp/clarin/webanno/ui/annotation/detail/AnnotationDetailEditorPanel.html
@@ -78,8 +78,7 @@
          -->
         <input wicket:id=forwardAnno tabindex="-1"></input>
       </form>
-      <div wicket:id="annotationFeatureForm" id="annotationFeatureForm" 
-          class="annotatation-detail-panel flex-content flex-v-container flex-gutter">
+      <div id="annotationFeatureForm" class="annotatation-detail-panel flex-content flex-v-container flex-gutter">
         <div wicket:id=deleteAnnotationDialog/>
         <div wicket:id=replaceAnnotationDialog/>
         <div wicket:id="layerContainer" class="card"/>
